+++ src/main/java/com/larsvansoest/runelite/clueitems/ui/StashUnitPanel.java
@@ -1,0 +1,10 @@
+package com.larsvansoest.runelite.clueitems.ui;
+
+import javax.swing.*;
+
+public class StashUnitPanel extends JPanel
+{
+	public StashUnitPanel(final Palette palette)
+	{
+	}
+}
